Refactor package of SQLDialectException
