+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/main/java/org/apache/shardingsphere/proxy/backend/handler/database/CreateDatabaseBackendHandler.java
@@ -19,7 +19,7 @@ package org.apache.shardingsphere.proxy.backend.handler.database;
 
 import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.proxy.backend.context.ProxyContext;
-import org.apache.shardingsphere.infra.exception.dialect.DBCreateExistsException;
+import org.apache.shardingsphere.infra.exception.dialect.database.DatabaseCreateExistsException;
 import org.apache.shardingsphere.proxy.backend.response.header.ResponseHeader;
 import org.apache.shardingsphere.proxy.backend.response.header.update.UpdateResponseHeader;
 import org.apache.shardingsphere.proxy.backend.handler.ProxyBackendHandler;
@@ -44,7 +44,7 @@ public final class CreateDatabaseBackendHandler implements ProxyBackendHandler {
     
     private void check(final CreateDatabaseStatement sqlStatement) {
         if (!sqlStatement.isIfNotExists() && ProxyContext.getInstance().databaseExists(sqlStatement.getDatabaseName())) {
-            throw new DBCreateExistsException(sqlStatement.getDatabaseName());
+            throw new DatabaseCreateExistsException(sqlStatement.getDatabaseName());
         }
     }
 }
