+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/main/java/org/apache/shardingsphere/proxy/backend/handler/distsql/ral/updatable/RefreshTableMetadataHandler.java
@@ -23,8 +23,8 @@ import org.apache.shardingsphere.infra.database.type.DatabaseTypeEngine;
 import org.apache.shardingsphere.infra.distsql.exception.DistSQLException;
 import org.apache.shardingsphere.mode.manager.ContextManager;
 import org.apache.shardingsphere.proxy.backend.context.ProxyContext;
-import org.apache.shardingsphere.infra.exception.dialect.NoDatabaseSelectedException;
-import org.apache.shardingsphere.infra.exception.dialect.UnknownDatabaseException;
+import org.apache.shardingsphere.infra.exception.dialect.database.NoDatabaseSelectedException;
+import org.apache.shardingsphere.infra.exception.dialect.database.UnknownDatabaseException;
 import org.apache.shardingsphere.proxy.backend.handler.distsql.ral.UpdatableRALBackendHandler;
 
 /**
